+++ example/sozlukdeneme/android/app/src/main/kotlin/com/example/sozlukdeneme/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.sozlukdeneme
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}