+++ lib/birlesim.dart
@@ -5,3 +5,5 @@ class Sozluk {
   var dictionary = a.Sozluk();
   var zemberek = b.Sozluk();
 }
+
+// Hey!! It's me. Again. Pffft. I'm CidQu btw.